Add private_notes field to Talk model
Adds a new TextField to the Talk model, but without adding it to the
user-visible side of things. Basically, I am relying on *not* adding it
to the TalkForm.Meta.fields set, which — as far as I understand — means
that the field is neither shown, nor considered if it were submitted.

This addresses #117, but does not solve it yet. The field is not yet
shown in detail view, nor is it available in the edit/change forms for
those that should be able to use it.