talks: Treat corresponding author like any other author
The corresponding author of a talk needs not be in the list of authors,
and this is a good separation, allowing conference organisers to create
talks for invited speakers etc.

However, post-creation, unless a staff member, the corresponding author
won't have edit access to the talk, which is herewith fixed.

Signed-off-by: martin f. krafft <madduck@madduck.net>