Don't use the configurable e-mail address for Debian SSO
Currently, users can SSO in, which creates an account. Then they change
their e-mail address (e.g. from an ugly alioth one, to their canonical
address) and the next time they SSO, they get a new account.

Rather, let's keep a wafer account associated with a Debian identity,
for life.

Storing this in the Registration group is a bit hacky, but this
information is useful to the registration team. It proves account
ownership.