Correctly determine if we're an admin
At some point in the past, "user" was the request user (or, at least, we
thought it was). It is now the same as object. So all the permission
checks that were based on user == object were returning true, when they
shouldn't have.

Instead, pass an explicit "can_edit" parameter to the template.