Add a last_updated field to schedule items
Various calendar exports want a last_updated value for indivual items,
so we make provision for that.

We update the last_updated time whenever the item is saved and whenever
the slots associated with the item are saved, so time changes are also
reflected.

Updating when the talk or page title is updated is a potential future
extension.