Replace markitup's inline JS to fix load order bug
See: https://github.com/zsiciarz/django-markitup/issues/25

django-markitup renders inline JS in each widget that configures the
widget. This JS requires jQuery and markitup.js to already be loaded,
which isn't the case if we're loading JS in the foot of the page.

So, override the template that generates that inline JS, with something
that generates a hidden tag containing only config. Then come back at
the end of the JS loading, and configure the widgets.

I'll do a similar PR upstream and hope that we can pull this out, later.