Add multisite support
All but the user models are now site-aware, i.e. by default they will
only load objects from the current site, and saving new objects will
automatically assign them to the current site.

This feature is opt-in, i.e. if you have a single conference and do
nothing, everything will just work as before.

A current limitation is that user management is global, and any
permissions given apply to all sites. I believe that addding support for
site-specific permission assignment is possible, and it's in my plans to
tackle that in future work.