templates: trim blocktrans strings
Adding `trimmed` to blocktrans compacts the strings in the PO files by
trimming whitespace at the beginning and end of the text. This way if
the blocktrans block gets e.g. its indentation changed, that will not
invalidate any translations.

This also makes the PO files a lot better to work with.