talk form: improve display of talk type and track descriptions
The formatting hides the description when the dropdown is closed, and
shows it in the dropdown itself, but less visible than the name. If JS
is turned off, the items will just present name and description.

This uses select2 directly instead of easy_select2, because easy_select2
does not allow to customize the appearance of the dropdown easily.