Stop using the Django CSRF token in the OAuth handshake
In 1.10, Django started to modify the CSRF token frequently to counter
BREACH. Simple string comparisons can no longer be used to validate the
token, and the token verification API isn't public.

https://code.djangoproject.com/ticket/20869
https://github.com/django/django/commit/5112e65ef2df1dbb95ff83026b6a962fb2688661

So, just use our own randomly generated secret, stored in the session.
It's only valid for a single handshake, and rotated on the next login
attempt.