schedule editor: hide/show items from bucket when (un)scheduling
i.e. when I drag and drop a talk from the "unassigned talks" bucket into
the schedule, it vanishes from the bucket. Conversely, when a talk is
deleted from the schedule it is displayed back in the bucket.

For this, we need to remove the duplicated id attribute in the "All
Talks" bucket.