Drop check for 'DATE-TIME' value field in 'DTSTAMP'
The paramaters field we checking in the test isn't actually required to
be present in the ical spec. The date we're interested in has to be a
DATE-TIME value by the spec, so the value field we're checking is just
optional rendundant information and icalendar no longer adds it by
default. We can make icalendar add it when generating the calendar, but
since it adds nothing useful and just increases the size of the ics
file, dropping this check seems the best approach.